requirements.txt
Pillow=11.1.0